Back out caa83f3d823e, Windows bustage.  r=bustage in a CLOSED TREE
UltraBlame original commit: d26e3ed73e410964cabc2c4711ca098b767800b9

+++ js/src/jsanalyze.cpp
@@ -518,7 +518,7 @@ ScriptAnalysis::analyzeLifetimes(JSContext *cx)
             
             for (unsigned i = 0; i < savedCount; i++) {
                 LifetimeVariable &var = *saved[i];
-                var.lifetime = alloc.new_<Lifetime>(offset, uint32_t(var.savedEnd), var.saved);
+                var.lifetime = alloc.new_<Lifetime>(offset, var.savedEnd, var.saved);
                 if (!var.lifetime) {
                     js_free(saved);
                     setOOM(cx);
@@ -610,8 +610,7 @@ ScriptAnalysis::analyzeLifetimes(JSContext *cx)
 
 
 
-                        var.lifetime =
-                            alloc.new_<Lifetime>(offset, uint32_t(var.savedEnd), var.saved);
+                        var.lifetime = alloc.new_<Lifetime>(offset, var.savedEnd, var.saved);
                         if (!var.lifetime) {
                             js_free(saved);
                             setOOM(cx);
@@ -675,8 +674,7 @@ ScriptAnalysis::addVariable(JSContext *cx, LifetimeVariable &var, unsigned offse
                 }
             }
         }
-        var.lifetime =
-            cx->typeLifoAlloc().new_<Lifetime>(offset, uint32_t(var.savedEnd), var.saved);
+        var.lifetime = cx->typeLifoAlloc().new_<Lifetime>(offset, var.savedEnd, var.saved);
         if (!var.lifetime) {
             setOOM(cx);
             return;
@@ -691,8 +689,7 @@ ScriptAnalysis::killVariable(JSContext *cx, LifetimeVariable &var, unsigned offs
 {
     if (!var.lifetime) {
         
-        Lifetime *lifetime =
-            cx->typeLifoAlloc().new_<Lifetime>(offset, uint32_t(var.savedEnd), var.saved);
+        Lifetime *lifetime = cx->typeLifoAlloc().new_<Lifetime>(offset, var.savedEnd, var.saved);
         if (!lifetime) {
             setOOM(cx);
             return;
